Revert "Added button groups with tab styles prop"
This reverts commit a3742a64d994bf5733456fb3c0ebfc3dea835d77.
