+++ packages/ui/core-components/src/lib/atoms/inputs/button-group/ButtonGroup.stories.svelte
@@ -136,39 +136,3 @@
 
 	Current Value: {$inputStore[args.name]}
 </Story>
-
-<Story name="Button Group Styled as Tabs" let:args>
-	<div class="mb-8">
-		<ButtonGroup
-			{...args}
-			preset="dates"
-			title="Default Button Style"
-			name="defaultStyle"
-			color="red"
-		/>
-	</div>
-	Current Value: {$inputStore['defaultStyle']}
-	<div class="mb-8 mt-4">
-		<ButtonGroup
-			{...args}
-			preset="dates"
-			display="tabs"
-			title="Buttons Styled as Tabs"
-			name="tabsStyle"
-		/>
-	</div>
-	Current Value: {$inputStore['tabsStyle']}
-</Story>
-
-<Story name="Hard Coded Entries with Tab Stylings" let:args>
-	<div class="mb-8">
-		<ButtonGroup {...args} display="tabs">
-			<ButtonGroupItem valueLabel="Option 1" value={1} />
-			<ButtonGroupItem valueLabel="Option 2" value={2} />
-			<ButtonGroupItem valueLabel="Option 3" value={3} />
-			<ButtonGroupItem valueLabel="Option 4" value={4} />
-		</ButtonGroup>
-	</div>
-
-	Current Value: {$inputStore[args.name]}
-</Story>
