+++ packages/ui/core-components/src/lib/atoms/inputs/button-group/ButtonGroup.svelte
@@ -6,7 +6,7 @@
 	import { presets, setButtonGroupContext } from './lib.js';
 	import { writable, readonly } from 'svelte/store';
 	import { INPUTS_CONTEXT_KEY } from '@evidence-dev/component-utilities/globalContexts';
-	import { getContext, setContext } from 'svelte';
+	import { getContext } from 'svelte';
 	import { buildReactiveInputQuery } from '@evidence-dev/component-utilities/buildQuery';
 	import ButtonGroupItem from './ButtonGroupItem.svelte';
 	import { page } from '$app/stores';
@@ -21,13 +21,6 @@
 
 	/** @type {keyof typeof presets | undefined} */
 	export let preset = undefined;
-
-	// for Tabs styling
-	export let display;
-
-	setContext('button-display', display);
-
-	export let color = 'hsla(207, 65%, 39%, 1)';
 
 	const inputs = getContext(INPUTS_CONTEXT_KEY);
 
@@ -56,28 +49,21 @@
 </script>
 
 <HiddenInPrint enabled={hideDuringPrint}>
-	<div
-		class={display === 'tabs' ? '' : 'inline-flex w-fit max-w-full flex-col mt-2 mb-4 ml-0 mr-2'}
-	>
+	<div class="inline-flex w-fit max-w-full flex-col mt-2 mb-4 ml-0 mr-2">
 		{#if title}
 			<span class="text-gray-900 text-sm block mb-1">{title}</span>
 		{/if}
-		<div
-			class={display === 'tabs'
-				? 'my-6 flex flex-wrap gap-x-1 gap-y-1'
-				: 'inline-flex rounded-md shadow-sm overflow-auto border no-scrollbar'}
-			role="group"
-		>
+		<div class="inline-flex rounded-md shadow-sm overflow-auto border no-scrollbar" role="group">
 			{#if preset}
 				{#if presets[preset]}
 					{#each presets[preset] as { value, valueLabel }}
-						<ButtonGroupItem {value} {valueLabel} {color} {display} />
+						<ButtonGroupItem {value} {valueLabel} />
 					{/each}
 				{:else}
 					<span class="text-red-500 font-bold text-sm">{preset} is not a valid preset</span>
 				{/if}
 			{:else}
-				<slot {display} />
+				<slot />
 				{#if hasQuery}
 					<QueryLoad data={query} let:loaded>
 						<svelte:fragment slot="skeleton">
@@ -85,7 +71,7 @@
 						</svelte:fragment>
 						<svelte:fragment>
 							{#each loaded as { label, value }}
-								<ButtonGroupItem {value} valueLabel={label} {color} {display} />
+								<ButtonGroupItem {value} valueLabel={label} />
 							{/each}
 						</svelte:fragment>
 					</QueryLoad>
